move logic to hook file
